Extract MinIO client functionality to separate project
